// libs/features/contributor/src/lib/shared/enums/days-left-labels.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum DaysLeftLabels {
  ZERO_DAYS = 'CONTRIBUTOR.DAY',
  ONE_DAY = 'CONTRIBUTOR.1-DAY',
  TWO_DAYS = 'CONTRIBUTOR.2-DAYS',
  THREE_TO_TEN_DAYS = 'CONTRIBUTOR.3-TO-10-DAYS',
  GRT_THAN_TEN_DAYS = 'CONTRIBUTOR.GRT-THAN-10-DAYS'
}