// libs/features/contributor/src/lib/shared/enums/reject-reasons.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

/* Person types enum */
export enum RejectReasonsEnum {
  INVALID_WAGE_DETAILS = 'INVALID_WAGE_DETAILS',
  INVALID_CONTRACT_DATE = 'INVALID_CONTRACT_DATE',
  INCORRECT_LABOR_RELATIONSHIP = 'INCORRECT_LABOR_RELATIONSHIP',
  OTHER_REASONS = 'OTHER_REASONS',
  INVALID_OCCUPATION ='INVALID_OCCUPATION'
}